scripts/requirements.txt
icalendar
pandas
caldav
cos-python-sdk-v5
tencentcloud-sdk-python